combining constraints, initialization, and a few methods
