+++ watertap/unit_models/reverse_osmosis_0D.py
@@ -18,7 +18,6 @@ from copy import deepcopy
 from pyomo.environ import (Var,
                            Set,
                            Param,
-                           Suffix,
                            NonNegativeReals,
                            NegativeReals,
                            Reference,
@@ -26,20 +25,12 @@ from pyomo.environ import (Var,
                            units as pyunits,
                            exp,
                            value)
-from pyomo.common.config import ConfigBlock, ConfigValue, In
 from pyomo.common.collections import ComponentSet
 # Import IDAES cores
 from idaes.core import (ControlVolume0DBlock,
-                        declare_process_block_class,
-                        MaterialBalanceType,
-                        EnergyBalanceType,
-                        MomentumBalanceType,
-                        UnitModelBlockData,
-                        useDefault)
-from idaes.core.util.config import is_physical_parameter_block
+                        declare_process_block_class)
 from idaes.core.util.exceptions import ConfigurationError
 from idaes.core.util import get_solver
-from idaes.core.util.tables import create_stream_table_dataframe
 import idaes.core.util.scaling as iscale
 from watertap.util.initialization import check_solve, check_dof
 from watertap.unit_models.membrane_base import (ConcentrationPolarizationType,
@@ -471,13 +462,6 @@ class ReverseOsmosisData(_MembraneBaseData):
                         sum(prop_io.flow_mass_phase_comp['Liq', j] for j in b.config.property_package.component_list)
                         * b.dh)
 
-            @self.Constraint(doc="Hydraulic diameter")  # eqn. 17 in Schock & Miquel, 1987
-            def eq_dh(b):
-                return (b.dh ==
-                        4 * b.spacer_porosity
-                        / (2 / b.channel_height
-                           + (1 - b.spacer_porosity) * 8 / b.channel_height))
-
         if self.config.pressure_change_type == PressureChangeType.fixed_per_unit_length:
             # Pressure change equation when dP/dx = user-specified constant,
             @self.Constraint(self.flowsheet().config.time,
@@ -819,20 +803,6 @@ class ReverseOsmosisData(_MembraneBaseData):
 
         return {"vars": var_dict}
 
-    def _get_stream_table_contents(self, time_point=0):
-        return create_stream_table_dataframe(
-            {
-                "Feed Inlet": self.inlet,
-                "Feed Outlet": self.retentate,
-                "Permeate Outlet": self.permeate,
-            },
-            time_point=time_point,
-        )
-
-    def get_costing(self, module=None, **kwargs):
-        self.costing = Block()
-        module.ReverseOsmosis_costing(self.costing, **kwargs)
-
     # permeate properties need to rescale solute values by 100
     def _rescale_permeate_variable(self, var, factor=100):
         if var not in self._permeate_scaled_properties:
@@ -863,33 +833,12 @@ class ReverseOsmosisData(_MembraneBaseData):
         iscale.set_scaling_factor(self.area, 1e-1)
 
         # setting scaling factors for variables
-        # these variables should have user input, if not there will be a warning
-        if iscale.get_scaling_factor(self.area) is None:
-            sf = iscale.get_scaling_factor(self.area, default=1, warning=True)
-            iscale.set_scaling_factor(self.area, sf)
 
         # these variables do not typically require user input,
         # will not override if the user does provide the scaling factor
-        if iscale.get_scaling_factor(self.A_comp) is None:
-            iscale.set_scaling_factor(self.A_comp, 1e12)
-
-        if iscale.get_scaling_factor(self.B_comp) is None:
-            iscale.set_scaling_factor(self.B_comp, 1e8)
-
         if iscale.get_scaling_factor(self.dens_solvent) is None:
             sf = iscale.get_scaling_factor(self.feed_side.properties_in[0].dens_mass_phase['Liq'])
             iscale.set_scaling_factor(self.dens_solvent, sf)
-
-        if iscale.get_scaling_factor(self.recovery_vol_phase) is None:
-            iscale.set_scaling_factor(self.recovery_vol_phase, 100.)
-
-        for (t, p, j), v in self.recovery_mass_phase_comp.items():
-            if j in self.config.property_package.solvent_set:
-                sf = 1
-            elif j in self.config.property_package.solute_set:
-                sf = 100
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
 
         for v in self.rejection_phase_comp.values():
             if iscale.get_scaling_factor(v) is None:
@@ -901,24 +850,24 @@ class ReverseOsmosisData(_MembraneBaseData):
                 iscale.set_scaling_factor(self.cp_modulus, sf)
 
         if hasattr(self, 'Kf_io'):
-            for t, x, j in self.Kf_io.keys():
-                if iscale.get_scaling_factor(self.Kf_io[t, x, j]) is None:
-                    iscale.set_scaling_factor(self.Kf_io[t, x, j], 1e5)
+            for v in self.Kf_io.values():
+                if iscale.get_scaling_factor(v) is None:
+                    iscale.set_scaling_factor(v, 1e5)
 
         if hasattr(self, 'N_Re_io'):
-            for t, x in self.N_Re_io.keys():
-                if iscale.get_scaling_factor(self.N_Re_io[t, x]) is None:
-                    iscale.set_scaling_factor(self.N_Re_io[t, x], 1e-2)
+            for v in self.N_Re_io.values():
+                if iscale.get_scaling_factor(v) is None:
+                    iscale.set_scaling_factor(v, 1e-2)
 
         if hasattr(self, 'N_Sc_io'):
-            for t, x in self.N_Sc_io.keys():
-                if iscale.get_scaling_factor(self.N_Sc_io[t, x]) is None:
-                    iscale.set_scaling_factor(self.N_Sc_io[t, x], 1e-2)
+            for v in self.N_Sc_io.values():
+                if iscale.get_scaling_factor(v) is None:
+                    iscale.set_scaling_factor(v, 1e-2)
 
         if hasattr(self, 'N_Sh_io'):
-            for t, x in self.N_Sh_io.keys():
-                if iscale.get_scaling_factor(self.N_Sh_io[t, x]) is None:
-                     iscale.set_scaling_factor(self.N_Sh_io[t, x], 1e-2)
+            for v in self.N_Sh_io.values():
+                if iscale.get_scaling_factor(v) is None:
+                     iscale.set_scaling_factor(v, 1e-2)
 
         if hasattr(self, 'length'):
             if iscale.get_scaling_factor(self.length) is None:
@@ -927,18 +876,6 @@ class ReverseOsmosisData(_MembraneBaseData):
         if hasattr(self, 'width'):
             if iscale.get_scaling_factor(self.width) is None:
                 iscale.set_scaling_factor(self.width, 1)
-
-        if hasattr(self, 'channel_height'):
-            if iscale.get_scaling_factor(self.channel_height) is None:
-                iscale.set_scaling_factor(self.channel_height, 1e3)
-
-        if hasattr(self, 'spacer_porosity'):
-            if iscale.get_scaling_factor(self.spacer_porosity) is None:
-                iscale.set_scaling_factor(self.spacer_porosity, 1)
-
-        if hasattr(self, 'dh'):
-            if iscale.get_scaling_factor(self.dh) is None:
-                iscale.set_scaling_factor(self.dh, 1e3)
 
         if hasattr(self, 'dP_dx'):
             for v in self.dP_dx.values():
@@ -1080,10 +1017,6 @@ class ReverseOsmosisData(_MembraneBaseData):
         if hasattr(self, 'eq_area'):
             sf = iscale.get_scaling_factor(self.area)
             iscale.constraint_scaling_transform(self.eq_area, sf)
-
-        if hasattr(self, 'eq_dh'):
-            sf = iscale.get_scaling_factor(self.dh)
-            iscale.constraint_scaling_transform(self.eq_dh, sf)
 
         if hasattr(self, 'eq_pressure_change'):
             for ind, c in self.eq_pressure_change.items():
